refactoring for move
