--- Duo.ConsoleTest/Duo.ConsoleTest.csproj
@@ -1,22 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0-windows10.0.19041.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Platform>x64</Platform>
-    <EnableWindowsTargeting>true</EnableWindowsTargeting> <!-- For SqlClient in macOS -->
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Duo.Core\Duo.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
