--- Duo.ConsoleTest/appsettings.json
@@ -1,7 +1,0 @@
-{
-    "LocalDataSource": "localhost",
-    "DataSource": "davidgatea.go.ro,1433",
-    "InitialCatalog": "duolingo_test",
-    "UserId": "REMOTE",
-    "Password": "duolingo123456789"
-}
