--- Duo.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Duo.Core;
-
-public class Class1
-{
-
-}
