--- Duo.Core/Data/Queries/User/GetById/query.sql
@@ -1,14 +1,0 @@
-CREATE PROCEDURE sp_GetUserById
-    @Id INT
-AS
-BEGIN
-    SET NOCOUNT ON;
-    
-    SELECT 
-        Id,
-        Username,
-        LastCompletedSectionId,
-        LastCompletedQuizId
-    FROM Users
-    WHERE Id = @Id;
-END
