--- Duo.Core/Data/Queries/User/GetByUsername/query.sql
@@ -1,14 +1,0 @@
-CREATE PROCEDURE sp_GetUserByUsername
-    @Username VARCHAR(100)
-AS
-BEGIN
-    SET NOCOUNT ON;
-    
-    SELECT 
-        Id,
-        Username,
-        LastCompletedSectionId,
-        LastCompletedQuizId
-    FROM Users
-    WHERE Username = @Username;
-END
