--- Duo.Core/Models/Difficulty.cs
@@ -1,8 +1,0 @@
-﻿namespace Duo.Models;
-
-public enum Difficulty
-{
-    Easy,
-    Normal,
-    Hard
-}
