--- Duo.Core/Models/Exercises/IExercise.cs
@@ -1,5 +1,0 @@
-﻿namespace Duo.Models.Exercises;
-
-interface IExercise
-{
-}
