--- Duo.Core/Models/Quizzes/IQuiz.cs
@@ -1,12 +1,0 @@
-﻿namespace Duo.Models.Quizzes;
-
-public interface IQuiz<T>
-{
-    bool AddExercise(T exercise);
-    bool IsValid();
-    double GetPassingThreshold();
-    public int GetNumberOfAnswersGiven();
-    public int GetNumberOfCorrectAnswers();
-    public void IncrementCorrectAnswers();
-    public void IncrementNumberOfAnswersGiven();
-}
