--- Duo.Core/Models/Roadmap/IRoadmap.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Duo.Models.Roadmap;
-
-interface IRoadmap<T>
-{
-    public void AddSection(T section);
-    public IEnumerable<T> GetAllSections();
-}
