--- Duo.Core/Services/UserService.cs
@@ -1,6 +1,0 @@
-namespace DefaultNamespace;
-
-public class UserService
-{
-    
-}
